backend/requirements.txt
Flask==3.1.0
gunicorn==23.0.0
lxml==5.3.0
arelle-release==2.36.40